Portfolio with Numbering
Portfolio with Numbering

+++ app/src/main/java/wit/cryptoexec/exchange/PortfolioInfo.java
@@ -1,10 +1,11 @@
 package wit.cryptoexec.exchange;
 
+import java.math.BigDecimal;
+
 public class PortfolioInfo {
     public String uuid;
     public String currency;
-    public Double balance;
-    public Double available;
+    public BigDecimal balance;
+    public BigDecimal available;
     public Double pending;
-    //public String cryptoAddress;
 }
